Refactor nova_signals.rs to use helper functions

+++ experiments/chimera-lang/src/vm/nova_biolum.rs
@@ -107,10 +107,10 @@ pub fn diffuse_light_color(vm: &mut ChimeraVM) {
 
             let inertia = vm.biome_grid[y][x].diffusion_inertia();
 
-            let mut r_acc: i64 = mr as i64 * my_light * inertia as i64;
-            let mut g_acc: i64 = mg as i64 * my_light * inertia as i64;
-            let mut b_acc: i64 = mb as i64 * my_light * inertia as i64;
-            let mut weight_acc: i64 = my_light * inertia as i64;
+            let mut r_acc: i64 = mr as i64 * my_light * inertia;
+            let mut g_acc: i64 = mg as i64 * my_light * inertia;
+            let mut b_acc: i64 = mb as i64 * my_light * inertia;
+            let mut weight_acc: i64 = my_light * inertia;
 
             for (dy, dx, mask) in NEIGHBOR_DIRECTIONS {
                 if (vm.membranes[y][x] & mask) != 0 {
